docker: add a Docker setup
Introduce a base Docker image and its companion docker-compose.yml
page to explore containerizing the project.

It's all, hopefully, straightforward but there's a small gotcha: the
web service (which runs the Rails server) and the JS service (which
runs the assets compilation) have to share the assets destination
path. That's what the `assets` volume does.

Everything else is pretty standard.